pyproject.toml
[project]
name = "ecommerce-agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-ollama>=0.3.8",
    "langgraph>=0.6.7",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.8",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
